hackhathon2/phase2/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
asyncpg==0.30.0
python-multipart==0.0.9
email-validator==2.1.0
